Add Azure publish properties
Add first name and last name to login page
Fix error when deleting user if he did not have profile picture

+++ Hybriotheca.Web/Controllers/UsersController.cs
@@ -291,7 +291,10 @@ namespace Hybriotheca.Web.Controllers
             {
                 await _userHelper.DeleteUserAsync(user);
 
-                await _blobHelper.DeleteBlobAsync(user.PhotoId.ToString(), "userphotos");
+                if (user.PhotoId != Guid.Empty)
+                {
+                    await _blobHelper.DeleteBlobAsync(user.PhotoId.ToString(), "userphotos");
+                }
             }
 
             return RedirectToAction(nameof(Index));
